requirements.txt
annotated-types==0.6.0
anthropic==0.25.7
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
arsenic==21.8
arxiv==2.1.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitarray==2.9.2
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
ckzg==1.0.2
click==8.1.7
cloudpickle==2.2.1
cohere==5.1.8
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
cytoolz==0.12.3
dask==2024.7.1
dask-expr==1.1.9
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dune_client==1.7.4
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
etherscan-python==2.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fastavro==1.9.4
feedparser==6.0.10
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.29.0
google_search_results==2.4.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
hexbytes==0.3.1
html2text==2024.2.26
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain
langchain-anthropic
langchain-cohere
langchain-community
langchain-mistralai
langchain-openai
langchain-text-splitters
langchainhub
langserve==0.0.48
langsmith==0.1.68
locket==1.0.0
lru-dict==1.2.0
lxml==5.2.2
markdown-it-py==3.0.0
marshmallow==3.20.2
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
metaphor-python==0.1.23
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
ndjson==0.3.1
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.10.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.3
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
selenium==4.19.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.12.0
snowflake-snowpark-python==1.20.0
sortedcontainers==2.4.0
soupsieve==2.5
spider-client==0.0.21
SQLAlchemy==2.0.25
sse-starlette==2.0.0
stack-data==0.6.3
starlette==0.35.1
structlog==20.2.0
swifter==1.4.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
tradingview-ta==3.3.0
traitlets==5.14.1
transformers==4.40.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240402
types-setuptools==70.1.0.20240627
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
urllib3==2.2.1
wcwidth==0.2.13
web3==6.19.0
websocket-client==1.8.0
websockets==10.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
